% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mar2.R
\name{estim_mar2_link}
\alias{estim_mar2_link}
\title{Detrend time series}
\usage{
estim_mar2_link(y)
}
\arguments{
\item{y}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
