% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mar2.R
\name{estim_mar_dep}
\alias{estim_mar_dep}
\title{Compute the parameter lambda between two MAR by the method of moments}
\usage{
estim_mar_dep(theta1, theta2, y)
}
\arguments{
\item{theta1}{a vector containing the MAR parameters returned by
\code{estim_mar2_link} associated with a first time series (eg residuals
related to global internal variability)}

\item{theta2}{a vector containing the MAR parameters returned by
\code{estim_mar2_link} associated with a second time series (eg residuals
related to local internal variability)}

\item{y}{a list of the two time series associated with
\code{theta1} and \code{theta2}}
}
\value{
a vector of parameters containing \code{theta1} and \code{theta2},
    and the parameter lambda.
}
\description{
Following equation 10 in Supplementary Material of Qasmi and Ribes (2021),
\code{estim_mar_dep} computes the parameter lambda by the method of moments
given two different MAR parameters.
}
\examples{

}
