% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mar2.R
\name{estim_mar_dep_full}
\alias{estim_mar_dep_full}
\title{Compute the parameter lambda between two MAR by maximum likelihood}
\usage{
estim_mar_dep_full(y)
}
\arguments{
\item{y}{a vector of two concatenated time series, typically corresponding to
global and local internal variability}
}
\value{
a vector containing the 9 MAR parameters fitted to \code{y}.
    \code{v1_slow} (\code{v1_fast}) is the variance of the white noise
    associated with the slow (fast) component of the first time series.
    \code{a1_ar1} (\code{a2_ar1}) is the coefficient at lag 1 associated with
    the slow (fast) component of the first time series. Similarly
    \code{v2_slow}, \code{v2_fast}, \code{a1_ar1}, \code{a2_ar1}) are the
    parameters associated with the second time series. \code{lambda} is the
    parameter linking the two time series.
}
\description{
\code{estim_mar_dep_full} computes all the parameters, including lambda,
relative to the mixture of the two autoregressive processes of order 1 (MAR,
see equation 10 in Supplementary Material of Qasmi and Ribes, 2021) to
capture fast and slow components within two given time series. Parameters are
 estimated by maximum likelihood via the \code{optim} function.
}
\examples{

}
