% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{full_names_to_std}
\alias{full_names_to_std}
\title{Extract model name from a pseudo observation chain of character}
\usage{
full_names_to_std(Models_fullnames, sep = "_")
}
\arguments{
\item{Models_fullnames}{a character, or a vector of character following the
pattern \code{model_member}}
}
\value{
a character, or a vector of character containing the \code{model}
    character.
}
\description{
For a given character folowing the pattern \code{model_member}, eg
\code{CanESM5_r1i1p1f1}, \code{full_names_to_std} returns the character
\code{model}, name, i.e. CanESM5.
}
\examples{

}
