% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fit.R
\name{hatm}
\alias{hatm}
\title{Calculate the 'smoothing matrix' of a spline smoothing}
\usage{
hatm(df, year)
}
\arguments{
\item{df}{a numeric, the equivalent degrees of freedom in the spline
smoothing.}

\item{year}{a vector of length n, giving the years of the time-series to
smooth.}
}
\value{
a nxn matrix, the smoothing matrix. Applying a spline smoothing is
equivalent to multiplying by this matrix.
}
\description{
Calculate the 'smoothing matrix' of a spline smoothing
}
