% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm_response.R
\name{hmodel}
\alias{hmodel}
\title{Calculate EBM model function}
\usage{
hmodel(FF, c, c0, lamb, gamm)
}
\arguments{
\item{FF}{A piControl time series}

\item{c}{A piControl time series}

\item{c0}{A piControl time series}

\item{lamb}{A piControl time series}

\item{gamm}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
