% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_List_function.R
\name{Impact_List_function}
\alias{Impact_List_function}
\title{Impact_List_function}
\usage{
Impact_List_function(x, ranking = T, ncores)
}
\arguments{
\item{x}{A numeric matrix or data.frame (sites × species).}

\item{ranking}{Logical; whether to rank impacts before SES calculation.}

\item{ncores}{Integer; number of cores for parallel computation.}
}
\value{
Impact list.#Note that the calculations performed in the function Impact_List_function should not be considered as estimates of community keystoneness. See main text.
}
\description{
Impact_List_function
}
