% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging}
\alias{kriging}
\title{Computes a set of realisations from a multivariate Gaussian distribution, and organizes the output as a X-array with "usual" dimensions, ie year, sample, forcing}
\usage{
kriging(x, SX, y, SY, H)
}
\arguments{
\item{x}{: a priori mean                background}

\item{SX}{: cov matrix of x                background error}

\item{y}{: observation                    obs}

\item{SY}{: cov matrix of y                observational error}

\item{H}{: observation operator        observation operator}
}
\value{
X a (ny,Nres+1,nf) array.
}
\description{
More detailed description
}
\examples{

}
