% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loadRData}
\alias{loadRData}
\title{Load a single object from a Rdata file into a specified variable}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{a character string giving the name of the file to load}
}
\description{
Load a single object from a Rdata file into a specified variable
}
\examples{

}
