% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{matrix_sqrt}
\alias{matrix_sqrt}
\title{Computes the square root of a matrix}
\usage{
matrix_sqrt(M)
}
\arguments{
\item{M}{a square matrix}
}
\value{
the square root of \code{M}
}
\description{
\code{matrix_sqrt} calculates the square root of a matrix by diagonalisation.
}
