% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mvgauss_to_Xarray}
\alias{mvgauss_to_Xarray}
\title{Computes an ensemble of realisations from a multivariate gaussian
distribution parameters}
\usage{
mvgauss_to_Xarray(mu, Sigma, Nres)
}
\arguments{
\item{mu}{a vector of time series corresponding to the mean of the
multivariate gaussian distribution. The names of each element must
correspond to the folowing pattern : \code{year_forcing}, eg
\code{1850_nat} for the mean natural response in 1850. The current
allowed forcings are \code{nat} and \code{all}. \code{mu} is tipically a
vector of concatenated time series relative to the response to the
natural and all external forcings for a given period.}

\item{Sigma}{a matrix of size \code{[length(mu),length(mu)]}
corresponding to the variance-covariance matrix of the multivariate
gaussian distribution}

\item{Nres}{the size of the wished sample}
}
\value{
returns a 3-D array of dimension
    \code{[Nres,length(year),length(forcing)]} where \code{year} is a time
    series of years and \code{forcing} is the type of forcing
}
\description{
For multiple types of external forcings (natural, all in the current
version), \code{mvgauss_to_Xarray} generates a sample of time series
corresponding to the response to each forcing from the gaussian parameters.
}
\examples{

}
