% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fit.R
\name{ones}
\alias{ones}
\title{Replace each element of a numeric vector by 1}
\usage{
ones(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of the same length as \code{x} but with each element equal
    to 1
}
\description{
Replace each element of a numeric vector by 1
}
\examples{

}
