% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fit.R
\name{pictl_detrend}
\alias{pictl_detrend}
\title{Remove trend from time series}
\usage{
pictl_detrend(x)
}
\arguments{
\item{x}{a vector of time series}
}
\value{
a vector from which the linear trend in \code{x} is removed
}
\description{
Remove trend from time series
}
\examples{

}
