% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cx}
\alias{plot_cx}
\title{Illustrate the observational constraint on covariate X}
\usage{
plot_cx(
  X,
  X_Cons,
  obs_x,
  ofile,
  ref_plot = NULL,
  sample = 0,
  event = NA,
  name_multi = "multi",
  ylim = NULL,
  add = F
)
}
\arguments{
\item{Xd}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
