% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scx}
\alias{plot_scx}
\title{Same as plot_cx, but inputs (X, X_Cons) include several scenarios, and are grouped into the same (big) array}
\usage{
plot_scx(
  X_Cons,
  obs_x,
  ofile,
  ref_plot = NULL,
  sample = 0,
  event = NA,
  name_multi = "multi",
  ylim = NULL
)
}
\arguments{
\item{Xd}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
