% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{prior2posterior}
\alias{prior2posterior}
\title{Calculate the prior distribution from X_fit, then applies the constrain, and returns a 4D array, including the "cons" and "uncons" vesrions of X.}
\usage{
prior2posterior(
  X_fit,
  Xo,
  Sigma_obs,
  Nres = NULL,
  centering_CX = T,
  ref_CX = "year_obs",
  S_mean = NULL,
  Sigma_mod = NULL,
  X_add = NULL
)
}
\arguments{
\item{x}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
