% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{prior2posterior_met}
\alias{prior2posterior_met}
\title{Derive the posterior distribution of time series from a prior given
observations (multiple metrics)
\code{prior2posterior_met} applies the same functions as
\code{prior2posterior} except that objects other than time series can be used
as inputs. In addition it takes as input the response to all external
forcings only, while \code{prior2posterior} can also consider the natural and
anthropogenic forcings.}
\usage{
prior2posterior_met(
  X_in,
  Xo,
  Sigma_obs,
  Nres = NULL,
  centering_CX = T,
  ref_CX = "year_obs",
  met_vec = 1
)
}
\arguments{
\item{X_in}{a 2-D array of dimension \code{[length(year), length(model)]},
which is a subset of an array returned by \code{x_fit}, containing the
best-estimated response to all external forcings and other metrics.}

\item{Xo}{a vector or a matrix. If a vector, \code{Xo} is a time series of
observations over a given period and/or observed metrics, and must have
names corresponding to the years of observations (eg \code{1850:2020}) or
to the metric name. If a matrix, lines correspond to the years of
observations and/or observed metrics, and columns to different type of
observations, which sample measurement uncertainty.}

\item{Sigma_obs}{a sum of a matrix sampling observed internal variability
and a matrix sampling error measurements in observations.}

\item{Nres}{the whished number of realisations in the posterior gaussian
sample}

\item{centering_CX}{a logical value indicating whether the constrained time
series must be in anomalies relative to a given period
(see \code{ref_CX})}

\item{ref_CX}{a vector containing the years corresponding to the reference
period if \code{centering_CX = TRUE}. If \code{ref_CX = "year_obs"} then
the reference is the whole period sampled by \code{year}.}

\item{met_vec}{an integer indicating the number of metrics.}
}
\value{
a list of two lists containing the parameters of the unconstrained
    (prior) and constrained (posterior) gaussian distributions for the
    response to all forcings. The first (second) list named
    \code{uncons} (\code{cons}) contains two other lists, namely \code{mean}
    and \code{var}. \code{mean} is a concatenation of time series
    corresponding to the mean of the distribution for the different forcings.
    The names of each element correspond to the folowing pattern :
    \code{year_all}, eg \code{1850_all} for the mean response in
    1850. \code{var} is the covariance matrix associated with \code{mu},
    sampling the model uncertainty.
}
\description{
Derive the posterior distribution of time series from a prior given
observations (multiple metrics)
\code{prior2posterior_met} applies the same functions as
\code{prior2posterior} except that objects other than time series can be used
as inputs. In addition it takes as input the response to all external
forcings only, while \code{prior2posterior} can also consider the natural and
anthropogenic forcings.
}
\examples{

}
