% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mar2.R
\name{Sigma_mar_dep}
\alias{Sigma_mar_dep}
\title{Compute the covariance matrix corresponding to a mixture of four
autoregressive (MAR) processes of order 1 and their link.}
\usage{
Sigma_mar_dep(theta, n)
}
\arguments{
\item{theta}{a vector containing the MAR parameters returned by
\code{estim_mar_dep} or \code{estim_mar_dep_full}}
}
\value{
a symmetric matrix corresponding to the combination of the two MAR
    models. The number of lines/columns in the covariance matrix, usually
    corresponds to the length of \code{y}.
}
\description{
\code{Sigma_mar_dep} takes the parameters returned by \code{estim_mar_dep} or
\code{estim_mar_dep_full} to calculate the matrix coefficients whose
formulation is given in equation 10 in Qasmi and Ribes (2021).
}
\examples{

}
