% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fit.R
\name{x_fit}
\alias{x_fit}
\title{Detrend time series}
\usage{
x_fit(Xd, Enat, Sigma, x_df, Nres, ant)
}
\arguments{
\item{Xd}{A piControl time series}

\item{Enat}{A piControl time series}

\item{Sigma}{A piControl time series}

\item{x_df}{A piControl time series}

\item{Nres}{A piControl time series}

\item{ant}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
