% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm_response.R
\name{ebm_response}
\alias{ebm_response}
\title{Calculate nat response for each set of params}
\usage{
ebm_response(FF, ebm_params, year, Nres)
}
\arguments{
\item{FF}{A piControl time series}

\item{ebm_params}{A piControl time series}

\item{year}{A piControl time series}

\item{Nres}{A piControl time series}
}
\description{
ANT NAT DECOMPOSITION PART
Calculate EBM responses --> Enat
}
\examples{

}
