% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging}
\alias{kriging}
\title{Apply the gaussian conditioning theorem to derive a posterior distribution
from a prior and observations}
\usage{
kriging(x, SX, y, SY, H)
}
\arguments{
\item{x}{a vector of time series corresponding to the mean of the prior}

\item{SX}{the covariance matrix corresponding to \code{x}. The number of
lines and columns must be equal to the length of \code{x}.}

\item{y}{a vector of time series corresponding to the observations}

\item{SY}{the covariance matrix corresponding to \code{y}. The number of
lines and columns must be equal to the length of \code{y}.}

\item{H}{a matrix corresponding to an observation operator. The number of
lines (columns) must be equal to the length of \code{y}(\code{x}).}
}
\value{
a list of two lists. The first list \code{mean} contains the mean of
    the posterior. The second list \code{var} contains the associated
    covariance matrix.
}
\description{
Apply the gaussian conditioning theorem to derive a posterior distribution
from a prior and observations
}
\examples{

}
