% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mvgauss_to_Xarray}
\alias{mvgauss_to_Xarray}
\title{Computes a set of realisations from a multivariate Gaussian distribution, and organizes the output as a X-array with "usual" dimensions, ie year, sample, forcing}
\usage{
mvgauss_to_Xarray(mu, Sigma, Nres)
}
\arguments{
\item{mu}{the mean of MV-Gauss distribution, a vector of size ns, mu needs appropriate names}

\item{Sigma}{the var-covariance matrix of MV-Gauss distribution, a (ns,ns) matrix}

\item{Nres}{the size of wished sample}
}
\value{
X a (ny,Nres+1,nf) array.
}
\description{
More detailed description
}
\examples{

}
