% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ones}
\alias{ones}
\title{Loading Rdata with a specific name}
\usage{
ones(x)
}
\arguments{
\item{x}{a}
}
\value{
As
}
\description{
creates an array (/vector/matrix) with only 1 [and appropriate names / dimnames]
}
\examples{

}
