% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fit.R
\name{pictl_detrend}
\alias{pictl_detrend}
\title{Detrend time series}
\usage{
pictl_detrend(x)
}
\arguments{
\item{x}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
