% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mar2.R
\name{Sigma_mar2}
\alias{Sigma_mar2}
\title{Detrend time series}
\usage{
Sigma_mar2(theta, n)
}
\arguments{
\item{y}{A piControl time series}
}
\value{
A detrended time series
}
\description{
More detailed description
}
\examples{

}
